loading: Delete `normpath` where possible
Unix path semantics have a design flaw, where if `a` is a symbolic
directory link, `x/a/..` refer's to the link target's parent directory
rather than `x/`. As a result, `normpath` is not in general safe
to run. We were doing this all over the place (either explicitly,
or implicitly in `abspath`) causing us to fail to load files if
their paths contained this pattern. Fix this by rm'ing normpath
in most places in loading and adding an `abspath` kwarg that skips
normalization. Where paths show up in printing, we can possibly
put some back after verifying that the path resolution result
does not change using `samefile`, but let's do that on a
case-by-case basis.